requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
distro==1.9.0
exceptiongroup==1.2.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-CORS==6.0.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.4
openai==1.35.13
outcome==1.3.0.post0
packaging==24.1
pandas==2.0.3
psycopg2-binary==2.9.6
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.45
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.4
Werkzeug==2.2.2
wikipedia==1.4.0
wsproto==1.2.0
